ISSUE #478 move ifc files to its own folder
