+++ bouncer/src/repo/manipulator/modelconvertor/import/ifcHelper/repo_ifc_utils_geometry.cpp
@@ -21,9 +21,9 @@
 
 #include "repo_ifc_utils_geometry.h"
 #include "repo_ifc_utils_constants.h"
-#include "../../../core/model/bson/repo_bson_factory.h"
+#include "../../../../core/model/bson/repo_bson_factory.h"
 #include <boost/filesystem.hpp>
-#include "repo_model_import_config_default_values.h"
+#include "../repo_model_import_config_default_values.h"
 using namespace repo::manipulator::modelconvertor;
 
 IFCUtilsGeometry::IFCUtilsGeometry(const std::string &file, const ModelImportConfig &settings) :
