+++ bouncer/src/repo/manipulator/modelconvertor/import/ifcHelper/repo_ifc_utils_parser.h
@@ -30,9 +30,9 @@
 #include <ifcparse/IfcParse.h>
 #include <ifcparse/IfcFile.h>
 
-#include "../../../core/model/bson/repo_node_material.h"
-#include "../../../core/model/bson/repo_node_mesh.h"
-#include "../../../core/model/collection/repo_scene.h"
+#include "../../../../core/model/bson/repo_node_material.h"
+#include "../../../../core/model/bson/repo_node_mesh.h"
+#include "../../../../core/model/collection/repo_scene.h"
 
 namespace repo {
 	namespace manipulator {
